+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Yii Framework 2 Change Log
 - Bug #3559: Use native support for batchInsert in SQLite versions >= 3.7.11 and avoid limitations of the fallback (cebe)
 - Bug #3564: Fixed the bug that primary key columns should not take default values from schema (qiangxue)
 - Bug #3567: Fixed the bug that smallint was treated as string for PostgreSQL (qiangxue)
+- Bug #3568: When the primary query sets `asArray`, it is not respected by the `via` relational query (qiangxue)
 - Bug #3578: Fixed postgreSQL column type detection, added missing types (MDMunir, cebe)
 - Bug #3583: Added typecast to auto value of primary key on insert of sql active record (cebe)
 - Bug #3591: Fixed incomplete obsolete filling in i18n `MessageController::saveMessagesToDb()` (advsm)